+++ db/sqlc/correction.sql.go
@@ -102,7 +102,7 @@ RETURNING id, problem, correct_info, email, status, request_date, verified_date,
 
 type UpdateCorrectionParams struct {
 	Status Statusrequest `json:"status"`
-	ID     int64         `json:"id"`
+	ID     int32         `json:"id"`
 }
 
 func (q *Queries) UpdateCorrection(ctx context.Context, arg UpdateCorrectionParams) (CorrectionForm, error) {
