+++ db/sqlc/school_rating.sql.go
@@ -21,8 +21,8 @@ INSERT INTO school_faculty_associations (
 `
 
 type CreateSchoolFacultyAssociationParams struct {
-	FacultyID int64 `json:"facultyID"`
-	SchoolID  int64 `json:"schoolID"`
+	FacultyID int32 `json:"facultyID"`
+	SchoolID  int32 `json:"schoolID"`
 }
 
 func (q *Queries) CreateSchoolFacultyAssociation(ctx context.Context, arg CreateSchoolFacultyAssociationParams) error {
@@ -52,7 +52,7 @@ INSERT INTO school_ratings (
 
 type CreateSchoolRatingParams struct {
 	UserID        string `json:"userID"`
-	SchoolID      int64  `json:"schoolID"`
+	SchoolID      int32  `json:"schoolID"`
 	Reputation    int16  `json:"reputation"`
 	Location      int16  `json:"location"`
 	Opportunities int16  `json:"opportunities"`
@@ -130,12 +130,12 @@ WHERE S.id = $1::bigint AND SR.id = $2::bigint
 `
 
 type GetSchoolRatingParams struct {
-	SchoolID       int64 `json:"schoolID"`
-	SchoolRatingID int64 `json:"schoolRatingID"`
+	SchoolID       int32 `json:"schoolID"`
+	SchoolRatingID int32 `json:"schoolRatingID"`
 }
 
 type GetSchoolRatingRow struct {
-	ID            int64  `json:"id"`
+	ID            int32  `json:"id"`
 	Reputation    int16  `json:"reputation"`
 	Location      int16  `json:"location"`
 	Opportunities int16  `json:"opportunities"`
@@ -196,13 +196,13 @@ OFFSET $3
 `
 
 type ListSchoolRatingsParams struct {
-	SchoolID int64 `json:"schoolID"`
+	SchoolID int32 `json:"schoolID"`
 	Limit    int32 `json:"limit"`
 	Offset   int32 `json:"offset"`
 }
 
 type ListSchoolRatingsRow struct {
-	ID            int64     `json:"id"`
+	ID            int32     `json:"id"`
 	Reputation    int16     `json:"reputation"`
 	Location      int16     `json:"location"`
 	Opportunities int16     `json:"opportunities"`
@@ -295,8 +295,8 @@ type UpdateSchoolRatingParams struct {
 	Review         sql.NullString `json:"review"`
 	UpVote         sql.NullInt32  `json:"upVote"`
 	DownVote       sql.NullInt32  `json:"downVote"`
-	SchoolRatingID int64          `json:"schoolRatingID"`
-	SchoolID       int64          `json:"schoolID"`
+	SchoolRatingID int32          `json:"schoolRatingID"`
+	SchoolID       int32          `json:"schoolID"`
 }
 
 func (q *Queries) UpdateSchoolRating(ctx context.Context, arg UpdateSchoolRatingParams) (SchoolRating, error) {
